Move `ImageCommentsPresentationAdapter` to production
