+++ EssentialApp/EssentialAppTests/ImageCommentsUIIntegrationTests.swift
@@ -12,34 +12,6 @@ import UIKit
 import EssentialFeed
 import EssentialFeediOS
 
-final class ImageCommentsPresentationAdapter: ImageCommentsViewControllerDelegate {
-	let imageLoader: ImageCommentsLoader
-	var presenter: ImageCommentsPresenter?
-	private var task: ImageCommmentsLoaderTask?
-	
-	init(imageLoader: ImageCommentsLoader) {
-		self.imageLoader = imageLoader
-	}
-	
-	func didRequestImageCommentsRefresh() {
-		presenter?.didStartLoadingComments()
-		
-		task = imageLoader.loadImageComments { [presenter] result in
-			switch result {
-			case let .success(comments):
-				presenter?.didFinishLoadingComments(with: comments)
-				
-			case let .failure(error):
-				presenter?.didFinishLoadingComments(with: error)
-			}
-		}
-	}
-	
-	func didCancelImageCommentsLoading() {
-		task?.cancel()
-	}
-	
-}
 
 final class ImageCommentsViewAdapter: ImageCommentsView {
 	private weak var controller: ImageCommentsViewController?
